Rename syncReqResp module
