+++ packages/lodestar/src/sync/reqResp/index.ts
@@ -1,2 +1,2 @@
 export * from "./interface";
-export * from "./syncReqResp";
+export * from "./reqResp";
